backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.10.0
joblib==1.4.2
scikit-learn==1.5.2
numpy==1.26.4
stripe==9.5.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
alembic==1.14.0
email-validator==2.2.0
PyJWT==2.10.1
bcrypt==4.2.1
httpx==0.28.1
transformers==4.46.3
torch==2.5.1
huggingface_hub==0.26.2
